CoxaController provides control entry point into the 6 x Dynamixel MX-28 coxa servos.

+++ include/l3xz/dynamixel/DynamixelController.h
@@ -31,7 +31,7 @@ namespace dynamixel
  * TYPEDEF
  **************************************************************************************/
 
-enum class Error : ssize_t
+enum class Error : int
 {
   None          =  0,
   AddParam      = -1,
